doon/requirements.txt
isort
django
psycopg2
djangorestframework
